+++ srcs/ft_math/tuple/tuple_tests2.c
@@ -4,6 +4,6 @@
 t_tuple	add_t(t_tuple a, t_tuple b);
 t_tuple	sub_t(t_tuple a, t_tuple b);
 t_vec	negate_v(t_vec v);
-t_vec	mult_v(t_vec v, float scalar);
-t_vec	div_v(t_vec v, float scalar);
+t_vec	mult_v(t_vec v, double scalar);
+t_vec	div_v(t_vec v, double scalar);
